moving docker/utils into container/docker
